Modified for Jenkins build
