+++ android/app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:fullBackupContent="false"
         android:hardwareAccelerated="true"
         tools:targetApi="honeycomb"
-        android:networkSecurityConfig="@xml/network_security_config"
+        tools:replace="android:fullBackupContent"
         tools:ignore="GoogleAppIndexingWarning">
 
         <!-- The main Corona activity that runs the Corona project. -->
@@ -27,7 +27,7 @@
             android:screenOrientation="portrait"
             android:configChanges="keyboard|keyboardHidden|navigation|screenSize|orientation"
             android:label="@string/app_name"
-            android:launchMode="singleTask"
+            android:launchMode="singleTop"
             android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
