+++ plugin_version.txt
@@ -1,1 +1,1 @@
-6.2.1
+6.3.0
