+++ plugins/2017.3184/iphone/metadata.lua
@@ -1,0 +1,17 @@
+local metadata =
+{
+	plugin =
+	{
+		format = 'staticLibrary',
+
+		-- This is the name without the 'lib' prefix.
+		-- In this case, the static library is called: libSTATIC_LIB_NAME.a
+		staticLibs = { 'plugin_vungle', 'z'}, 
+
+		frameworks = {'AVFoundation', 'CFNetwork', 'CoreGraphics', 'AudioToolbox', 'Accounts', 'AdSupport', 'CoreMedia',
+				   'Foundation', 'MediaPlayer', 'QuartzCore', 'SystemConfiguration', 'StoreKit', 'Photos', 'AssetsLibrary'},
+		frameworksOptional = {'CoreLocation', 'WebKit'},
+	},
+}
+
+return metadata
