+++ plugins/2018.3326/win32-sim/plugin_vungle.lua
@@ -1,0 +1,16 @@
+local Library = require "CoronaLibrary"
+
+-- Create stub library for simulator
+local lib = Library:new{ name='plugin.vungle', publisherId='com.vungle' }
+
+-- Default implementations
+local function defaultFunction()
+    print( "WARNING: The '" .. lib.name .. "' library is not available on this platform." )
+end
+
+lib.init= defaultFunction
+lib.requestToken= defaultFunction
+lib.allowsApplePay= defaultFunction
+
+-- Return an instance
+return lib
